+++ src/collective/volto/blocksfield/adapters/searchabletext_indexers.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from plone.app.dexterity.textindexer.converters import (
-DefaultDexterityTextIndexFieldConverter,
-)
-from plone.app.dexterity.textindexer.interfaces import (
-    IDexterityTextIndexFieldConverter,
-)
 from collective.volto.blocksfield.interfaces import IBlocksField
+from plone.app.dexterity.textindexer.converters import (
+    DefaultDexterityTextIndexFieldConverter,
+)
+from plone.app.dexterity.textindexer.interfaces import IDexterityTextIndexFieldConverter
 from plone.dexterity.interfaces import IDexterityContent
 from plone.restapi.interfaces import IBlockSearchableText
 from z3c.form.interfaces import IWidget
@@ -38,7 +36,6 @@ class BlocksFieldConverter(DefaultDexterityTextIndexFieldConverter):
                 IBlockSearchableText,
                 name=block_type,
             )
-
             if adapter is not None:
                 text = adapter(block)
                 if text: