+++ CHANGES.rst
@@ -1,15 +1,19 @@
 Changelog
 =========
 
+2.0.1 (unreleased)
+------------------
+- Add serializers and deserializer for slate blocks.
+  [cekk]
+- Add indexer for slate blocks.
+  [cekk]
 
-2.0.0 (unreleased)
+2.0.0 (2023-07-03)
 ------------------
 
 - Drop support for Plone 5.2 and collective.dexteritytextindexer; in Plone6
   we use core functions in plone.app.dexterity.textindexer
   [lucabel]
-- Add serializers and deserializer for slate blocks.
-  [cekk]
 
 1.0.3 (2022-05-31)
 ------------------